Adjusting models for BenchmarkInput
