+++ mcs_benchmark_data/models/benchmark_concept.py
@@ -1,24 +1,22 @@
 from dataclasses import dataclass
 from dataclasses_json import LetterCase, dataclass_json
 
-from rdflib import Graph, URIRef
+from rdflib import Graph
 from rdflib.resource import Resource
-from mcs_benchmark_data.namespace import MCS, SCHEMA
+from mcs_benchmark_data.namespace import SCHEMA
 
-from mcs_benchmark_data._model import _Model
+from mcs_benchmark_data.models.benchmark_input import BenchmarkInput
 
 
 @dataclass_json(letter_case=LetterCase.CAMEL)
 @dataclass(frozen=True)
-class BenchmarkConcept(_Model):
+class BenchmarkConcept(BenchmarkInput):
     """The ConceptNet concept which the question was created from (i.e. electricity)"""
 
-    antecedent_uri: URIRef
     concept: str
 
     def to_rdf(self, *, graph: Graph) -> Resource:
-        resource = _Model.to_rdf(self, graph=graph)
-        graph.add((self.antecedent_uri, MCS.BenchmarkConcept, self.uri))
+        resource = super().to_rdf(graph=graph)
         resource.add(SCHEMA.text, self._quote_rdf_literal(self.concept))
 
         return resource
