+++ mcs_benchmark_data/models/benchmark_prompt.py
@@ -1,11 +1,11 @@
 from dataclasses import dataclass
 
 from dataclasses_json import LetterCase, dataclass_json
-from rdflib import Graph, URIRef
+from rdflib import Graph
 from rdflib.resource import Resource
 
 from mcs_benchmark_data._model import _Model
-from mcs_benchmark_data.namespace import MCS, SCHEMA
+from mcs_benchmark_data.namespace import SCHEMA
 
 
 @dataclass_json(letter_case=LetterCase.CAMEL)
